4/study-hybrid-search-gcp/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Cloud Run Service — FastAPI prediction API"
requires-python = ">=3.12"
dependencies = [
    "common",
    "cachetools>=5.3",
    "fastapi>=0.115",
    "google-auth>=2.36",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "jinja2>=3.1",
    "lightgbm>=4.5",
    "numpy>=2.0",
    "sentence-transformers>=3.2",
    "google-cloud-pubsub>=2.23",
    "google-cloud-run>=0.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
